@@ -22,30 +22,30 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@frequency-chain/api-augment": "file:../js/api-augment/dist/frequency-chain-api-augment-0.0.0.tgz",
-    "@helia/unixfs": "^5.0.0",
-    "@noble/curves": "^1.8.2",
-    "@polkadot-api/merkleize-metadata": "^1.1.14",
-    "@polkadot/api": "15.9.2",
-    "@polkadot/types": "15.9.2",
+    "@helia/unixfs": "^5.0.2",
+    "@noble/curves": "^1.9.1",
+    "@polkadot-api/merkleize-metadata": "^1.1.15",
+    "@polkadot/api": "15.10.2",
+    "@polkadot/types": "15.10.2",
     "@polkadot/util": "13.4.4",
     "ethers": "^6.14.0",
-    "helia": "^5.3.0",
-    "multiformats": "^13.3.2",
+    "helia": "^5.4.1",
+    "multiformats": "^13.3.3",
     "rxjs": "^7.8.2",
     "workerpool": "^9.2.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.25.0",
+    "@eslint/js": "^9.26.0",
     "@types/mocha": "^10.0.10",
-    "eslint": "^9.25.0",
-    "eslint-plugin-mocha": "^10.5.0",
-    "glob": "^11.0.1",
-    "globals": "^16.0.0",
-    "mocha": "^11.1.0",
+    "eslint": "^9.26.0",
+    "eslint-plugin-mocha": "^11.0.0",
+    "glob": "^11.0.2",
+    "globals": "^16.1.0",
+    "mocha": "^11.2.2",
     "prettier": "^3.5.3",
     "sinon": "^20.0.0",
-    "tsx": "^4.19.3",
+    "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.30.1"
+    "typescript-eslint": "^8.32.1"
   }
 }
