@@ -34,21 +34,21 @@
   "author": "frequency-chain",
   "license": "Apache-2.0",
   "dependencies": {
-    "@polkadot/api": "^15.9.2",
-    "@polkadot/rpc-provider": "^15.9.2",
-    "@polkadot/types": "^15.9.2",
-    "globals": "^16.0.0"
+    "@polkadot/api": "^15.10.2",
+    "@polkadot/rpc-provider": "^15.10.2",
+    "@polkadot/types": "^15.10.2",
+    "globals": "^16.1.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.25.0",
-    "@polkadot/typegen": "^15.9.2",
+    "@eslint/js": "^9.26.0",
+    "@polkadot/typegen": "^15.10.2",
     "@types/mocha": "^10.0.10",
-    "eslint": "^9.25.0",
-    "eslint-plugin-mocha": "^10.5.0",
-    "mocha": "11.1.0",
+    "eslint": "^9.26.0",
+    "eslint-plugin-mocha": "^11.0.0",
+    "mocha": "11.2.2",
     "prettier": "^3.5.3",
-    "tsx": "^4.19.3",
+    "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.30.1"
+    "typescript-eslint": "^8.32.1"
   }
 }
