@@ -160,7 +160,13 @@ describe('Capacity Transaction Failures', function () {
       // from the transaction pool.
       it('fails to pay for a transaction with empty capacity', async function () {
         const noCapacityKeys = createKeys('noCapacityKeys');
-        const noCapacityProvider = await createMsaAndProvider(fundingSource, noCapacityKeys, 'NoCapProvider');
+        const noCapacityProvider = await createMsaAndProvider(
+          fundingSource,
+          noCapacityKeys,
+          'NoCapProvider',
+          undefined,
+          false
+        );
 
         const delegatorKeys = createKeys('delegatorKeys');
 
