@@ -181,7 +181,8 @@ describe('Capacity Transactions', function () {
 
           const { eventMap } = await call.payWithCapacity();
           assertEvent(eventMap, 'capacity.CapacityWithdrawn');
-          assertEvent(eventMap, 'messages.MessagesInBlock');
+          // messages.MessagesInBlock in block might not be on this transaction if there are others
+          assertEvent(eventMap, 'system.ExtrinsicSuccess');
         });
 
         it('successfully pays with Capacity for eligible transaction - addOnchainMessage', async function () {
@@ -190,7 +191,8 @@ describe('Capacity Transactions', function () {
           const call = ExtrinsicHelper.addOnChainMessage(capacityKeys, dummySchemaId, '0xdeadbeef');
           const { eventMap } = await call.payWithCapacity();
           assertEvent(eventMap, 'capacity.CapacityWithdrawn');
-          assertEvent(eventMap, 'messages.MessagesInBlock');
+          // messages.MessagesInBlock in block might not be on this transaction if there are others
+          assertEvent(eventMap, 'system.ExtrinsicSuccess');
           const get = await ExtrinsicHelper.apiPromise.rpc.messages.getBySchemaId(dummySchemaId, {
             from_block: starting_block,
             from_index: 0,
