@@ -3,27 +3,13 @@
 import eslint from '@eslint/js';
 import globals from 'globals';
 import tseslint from 'typescript-eslint';
-import mocha from 'eslint-plugin-mocha';
-
-// Needed for eslint 9
-const mochaConfig = [
-  {
-    name: 'mocha/recommended',
-    languageOptions: {
-      globals: globals.mocha,
-    },
-    plugins: {
-      mocha,
-    },
-    rules: mocha.configs.flat.recommended.rules,
-  },
-];
+import mochaPlugin from 'eslint-plugin-mocha';
 
 export default tseslint.config(
   eslint.configs.recommended,
+  mochaPlugin.configs.recommended,
   ...tseslint.configs.strict,
   ...tseslint.configs.stylistic,
-  ...mochaConfig,
   {
     ignores: ['dist/'],
   },
