@@ -33,36 +33,42 @@ describe('📗 Stateful Pallet Storage Ethereum', function () {
   let ethereumDelegatorKeys: KeyringPair;
 
   before(async function () {
-    // Create a provider. This provider will NOT be granted delegations;
-    // methods requiring a payload signature do not require a delegation
-    [undelegatedProviderKeys, undelegatedProviderId] = await createProviderKeysAndId(fundingSource, 2n * DOLLARS);
+    // All the setup
+    [
+      // Create a provider. This provider will NOT be granted delegations;
+      // methods requiring a payload signature do not require a delegation
+      [undelegatedProviderKeys, undelegatedProviderId],
+      // Create a provider for the MSA, the provider will be used to grant delegation
+      [delegatedProviderKeys, delegatedProviderId],
+      // Create a schema for Itemized PayloadLocation
+      itemizedSchemaId,
+      // Create a schema for Paginated PayloadLocation
+      paginatedSchemaId,
+    ] = await Promise.all([
+      createProviderKeysAndId(fundingSource, 2n * DOLLARS),
+      createProviderKeysAndId(fundingSource, 2n * DOLLARS),
+      ExtrinsicHelper.getOrCreateSchemaV3(
+        fundingSource,
+        AVRO_CHAT_MESSAGE,
+        'AvroBinary',
+        'Itemized',
+        ['AppendOnly', 'SignatureRequired'],
+        'test.ItemizedSignatureRequired'
+      ),
+      ExtrinsicHelper.getOrCreateSchemaV3(
+        fundingSource,
+        AVRO_CHAT_MESSAGE,
+        'AvroBinary',
+        'Paginated',
+        ['SignatureRequired'],
+        'test.PaginatedSignatureRequired'
+      ),
+    ]);
     assert.notEqual(undelegatedProviderId, undefined, 'setup should populate undelegatedProviderId');
     assert.notEqual(undelegatedProviderKeys, undefined, 'setup should populate undelegatedProviderKeys');
 
-    // Create a provider for the MSA, the provider will be used to grant delegation
-    [delegatedProviderKeys, delegatedProviderId] = await createProviderKeysAndId(fundingSource, 2n * DOLLARS);
     assert.notEqual(delegatedProviderId, undefined, 'setup should populate delegatedProviderId');
     assert.notEqual(delegatedProviderKeys, undefined, 'setup should populate delegatedProviderKeys');
-
-    // Create a schema for Itemized PayloadLocation
-    itemizedSchemaId = await ExtrinsicHelper.getOrCreateSchemaV3(
-      undelegatedProviderKeys,
-      AVRO_CHAT_MESSAGE,
-      'AvroBinary',
-      'Itemized',
-      ['AppendOnly', 'SignatureRequired'],
-      'test.ItemizedSignatureRequired'
-    );
-
-    // Create a schema for Paginated PayloadLocation
-    paginatedSchemaId = await ExtrinsicHelper.getOrCreateSchemaV3(
-      undelegatedProviderKeys,
-      AVRO_CHAT_MESSAGE,
-      'AvroBinary',
-      'Paginated',
-      ['SignatureRequired'],
-      'test.PaginatedSignatureRequired'
-    );
 
     // Create a MSA for the delegator
     [ethereumDelegatorKeys, msa_id] = await createDelegatorAndDelegation(
@@ -72,9 +78,10 @@ describe('📗 Stateful Pallet Storage Ethereum', function () {
       delegatedProviderKeys,
       'ethereum'
     );
-    console.log('after createDelegatorAndDelegation');
     assert.notEqual(ethereumDelegatorKeys, undefined, 'setup should populate delegator_key');
     assert.notEqual(msa_id, undefined, 'setup should populate msa_id');
+    // Make sure we are finalized before all the tests
+    await ExtrinsicHelper.waitForFinalization();
   });
 
   describe('Itemized With Signature Storage Tests', function () {
